moved VariantAnnotation persistence and transformation methods into a custom VariantAnnotationRepository class
