+++ annotation/src/main/java/org/cbioportal/genome_nexus/annotation/web/AnnotationController.java
@@ -32,14 +32,10 @@
 
 package org.cbioportal.genome_nexus.annotation.web;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mongodb.DBObject;
-import com.mongodb.util.JSON;
 import io.swagger.annotations.*;
 import org.cbioportal.genome_nexus.annotation.domain.*;
 import org.cbioportal.genome_nexus.annotation.service.*;
 
-import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.client.HttpClientErrorException;
@@ -58,16 +54,13 @@ public class AnnotationController
 {
     private final VariantAnnotationService variantAnnotationService;
     private final VariantAnnotationRepository variantAnnotationRepository;
-    private final MongoTemplate mongoTemplate;
 
     @Autowired
     public AnnotationController(VariantAnnotationService variantAnnotationService,
-                                VariantAnnotationRepository variantAnnotationRepository,
-                                MongoTemplate mongoTemplate)
+                                VariantAnnotationRepository variantAnnotationRepository)
     {
         this.variantAnnotationService = variantAnnotationService;
         this.variantAnnotationRepository = variantAnnotationRepository;
-        this.mongoTemplate = mongoTemplate;
     }
 
     @ApiOperation(value = "getVariantAnnotation",
@@ -140,10 +133,10 @@ public class AnnotationController
                 // construct a VariantAnnotation instance to return:
                 // this does not contain all the information obtained from the web service
                 // only the fields mapped to the VariantAnnotation model will be returned
-                variantAnnotation = mapAnnotationJson(variant, annotationJSON);
+                variantAnnotation = variantAnnotationRepository.mapAnnotationJson(variant, annotationJSON);
 
                 // save everything to the cache as a properly parsed JSON
-                saveToDB(variant, annotationJSON);
+                variantAnnotationRepository.saveAnnotationJson(variant, annotationJSON);
             }
             catch (HttpClientErrorException e) {
                 // in case of web service error, do not terminate the whole process.
@@ -158,65 +151,5 @@ public class AnnotationController
         }
 
         return variantAnnotation;
-    }
-
-    // TODO move this method into VariantAnnotationRepositoryImpl class if possible
-	/**
-     * Parses and saves the entire content of the annotation JSON object to the database.
-     *
-     * @param variant           variant key (used as an id)
-     * @param annotationJSON    raw annotation JSON (obtained from the service)
-     */
-    private void saveToDB(String variant, String annotationJSON)
-    {
-        // parse the given annotation JSON string to get a proper object
-        DBObject dbObject = (DBObject)JSON.parse(removeArrayBrackets(annotationJSON));
-
-        // update the _id field to the given variable
-        dbObject.put("_id", variant);
-
-        // save the object into the correct repository
-        this.mongoTemplate.save(dbObject, "vep.annotation");
-    }
-
-	/**
-     * Maps the given raw annotation JSON string onto a VariantAnnotation instance.
-     *
-     * @param variant           variant key
-     * @param annotationJSON    raw annotation JSON string
-     * @return a VariantAnnotation instance
-     * @throws IOException
-     */
-    private VariantAnnotation mapAnnotationJson(String variant, String annotationJSON) throws IOException
-    {
-        String toSerialize = removeArrayBrackets(annotationJSON);
-
-        VariantAnnotation vepVariantAnnotation;
-        ObjectMapper mapper = new ObjectMapper();
-
-        // map annotation string onto VariantAnnotation instance
-        vepVariantAnnotation = mapper.readValue(toSerialize, VariantAnnotation.class);
-        // include original variant value too
-        vepVariantAnnotation.setVariant(variant);
-        //vepVariantAnnotation.setAnnotationJSON(annotationJSON);
-
-        return vepVariantAnnotation;
-    }
-
-    private String removeArrayBrackets(String annotationJSON)
-    {
-        String stripped = annotationJSON;
-
-        // TODO this is not safe. a better way: parse and if array then get the first element
-        // if annotation JSON is actually an array, assume it is an array of size 1,
-        // and convert it to a single object by removing array brackets
-        if (annotationJSON.trim().startsWith("["))
-        {
-            int start = annotationJSON.indexOf("[");
-            int end = annotationJSON.lastIndexOf("]");
-            stripped = annotationJSON.substring(start+1, end);
-        }
-
-        return stripped;
     }
 }
